% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expose_cells.R
\name{expose_cells}
\alias{expose_cells}
\title{Expose a seed population of cells to a mutagen (or control)}
\usage{
expose_cells(
  population = data.frame(),
  n_div = 3,
  mut_freq = 0.1,
  mut_process = ""
)
}
\arguments{
\item{population}{A data frame representing your initial cell popultion.
This object should be an output from `generate_population`.}

\item{n_div}{The number of cell divisions to simulate.}

\item{mut_freq}{The frequency of mutation to use in simulations.}

\item{mut_process}{The overall mutation spectrum to simulate, provided as...?}
}
\description{
Simulates exposure of a cell population (provided as a dataframe which
should be created using the `generate_population` function). Your initial
cell population will expand over `n_div`, a given number of cell divisions,
using `mut_freq` and `mut_process` as the frequency of mutations and
spectrum to generate, respectively.
}
