% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_population.R
\name{generate_population}
\alias{generate_population}
\title{Generate a seed population of cells with mutations}
\usage{
generate_population(
  n_div = 3,
  n_start = 1,
  ref = "hg38",
  mut_freq = 0.1,
  mut_process = ""
)
}
\arguments{
\item{n_div}{The number of cell divisions to simulate.}

\item{n_start}{The number of starting cells in the population.}

\item{ref}{Reference genome; hg38, mm10, or a path to a FASTA file}

\item{mut_freq}{The frequency of mutation to use in simulations.}

\item{mut_process}{The overall mutation spectrum to simulate, provided as...?}
}
\description{
Create a digital "tissue", and return all the mutations
that have arisen within that "tissue". The purpose is to create a mock cell
population, over `n_div` rounds of cell division, starting with `n_start`
cells, using a mutation frequency, `mut_freq` and mutation process,
`mut_process`, across a reference DNA sequence, `ref`. At the end, you will
end up with a starting population of cells that "grew" using these specific
parameters. Then, this tissue can be mock "exposed" (or "expanded") to
simulate a cellular exposure to a mutagen (or control samples).
}
